Fixed UPDATE and DATE as well as a few other bugs
Updates all dates and info to the database.

since the dates still default to mm/dd/yyyy I changed it so that it will ignore those requests that return date as '0000-00-00' or null or a previous date value and instead assign the $scope to the value before the eit button was hit.

additionally I had to create a new PHP file for the test table so that the SQL would read correctly. It was adapted from the import statement and might have future problems, but none so far that I noticed

TODO - page needs to refresh component after update. Still need to format update button. Would be nice to have dates default to the value on the View tab, but it still functions correctly currently.