updated move functionality
implemented the CSS phelps wants on our files, but it still needs work. The table headers will not show and the table is not centered.
The move functions work, but don't reliably delete the user from the database after being moved. 
The move functions run into issues when there are two applicants with the same applicantID (in different tables). We need to find a way to keep those unique across tables. 

I've been working on the mailing function, but it was messing everything up so it is commented out on the files.